Remove `Extractor` trait.
Move `Extractor` to top level module.
