+++ rust/src/openvasd/container_image_scanner/scheduling/scanner.rs
@@ -210,17 +210,16 @@ async fn is_digest_excluded(
         false
     }
 }
-async fn download_and_extract_image<'a, E, R>(
+async fn download_and_extract_image<'a, R>(
     config: Arc<Config>,
     pool: &DataBase,
     registry: &'a super::InitializedRegistry<'a, R>,
     image: Image,
-) -> Result<(Digest, E, Vec<Benched>), ScannerError>
+) -> Result<(Digest, Extractor, Vec<Benched>), ScannerError>
 where
-    E: Extractor + Send + Sync,
     R: Registry + Send + Sync,
 {
-    let mut extractor = E::initialize(config.clone(), registry.id.clone()).await?;
+    let mut extractor = Extractor::initialize(config.clone(), registry.id.clone()).await?;
     let mut results = Vec::new();
     let mut digest = None;
 
@@ -263,14 +262,13 @@ where
     Ok((digest.unwrap_or_default(), extractor, results))
 }
 
-async fn retry_download_and_extract_image<'a, E, R>(
+async fn retry_download_and_extract_image<'a, R>(
     config: Arc<Config>,
     pool: &DataBase,
     registry: &'a super::InitializedRegistry<'a, R>,
     image: &Image,
-) -> Result<(Image, E), ScannerError>
+) -> Result<(Image, Extractor), ScannerError>
 where
-    E: Extractor + Send + Sync,
     R: Registry + Send + Sync,
 {
     // alternatively set back to pending and store retry amount alongside the image
@@ -293,14 +291,13 @@ where
     }
 }
 
-pub async fn scan_image<'a, E, R, T>(
+pub async fn scan_image<'a, R, T>(
     config: Arc<Config>,
     pool: DataBase,
     products: Arc<RwLock<Notus>>,
     registry: &'a super::InitializedRegistry<'a, R>,
 ) -> Result<(), Vec<ScannerError>>
 where
-    E: Extractor + Send + Sync,
     R: Registry + Send + Sync,
     T: ToNotus,
 {
@@ -311,7 +308,7 @@ where
         .map_err(|e| vec![ScannerError::from(e)])?;
 
     let (digest, locator_per_arch) =
-        retry_download_and_extract_image::<E, _>(config, &pool, registry, &image)
+        retry_download_and_extract_image(config, &pool, registry, &image)
             .await
             .map_err(|e| vec![e])?;
     let locator_per_arch = locator_per_arch.locator().await;
